fix: remove callback from createWorkflow result
it was needed before for making request payload and return types of workflows typesafe. now that we have routeFunction in the result, we don't need the callback.
