+++ src/serve/serve-many.test.ts
@@ -1,7 +1,5 @@
 import { describe, expect, test } from "bun:test";
 import { createWorkflow, serveMany } from "../../platforms/nextjs";
-import { Telemetry } from "../types";
-import { createInvokeCallback } from "./serve-many";
 import { WorkflowContext } from "../context";
 import { Client } from "@upstash/qstash";
 import { getRequest, MOCK_QSTASH_SERVER_URL, mockQStashServer, WORKFLOW_ENDPOINT } from "../test-utils";
@@ -27,88 +25,6 @@ describe("serveMany", () => {
       const throws = async () => await handler(request)
 
       expect(throws).toThrow("No workflows in serveMany found for 'workflow-two'")
-    })
-  })
-
-  describe("createInvokeCallback", () => {
-    test("should call create invoke", async () => {
-      const token = nanoid();
-
-      const telemetry: Telemetry = {
-        sdk: "sdk",
-        framework: "framework",
-        runtime: "runtime",
-      }
-      const callback = createInvokeCallback(telemetry)
-      const workflowId = "some-workflow-id"
-
-      await mockQStashServer({
-        execute: async () => {
-
-          await callback({
-            body: "some-body",
-            workflow: {
-              workflowId,
-              callback,
-              options: {}
-            },
-            headers: { "custom": "custom-header-value" },
-            retries: 6,
-            workflowRunId: "some-run-id",
-          }, {
-            stepId: 4,
-            concurrent: 1,
-            stepName: "invoke-step",
-            stepType: "Invoke",
-          }, new WorkflowContext({
-            headers: new Headers({ "original": "original-headers-value" }) as Headers,
-            initialPayload: "initial-payload",
-            qstashClient: new Client({ baseUrl: MOCK_QSTASH_SERVER_URL, token }),
-            steps: [],
-            url: `${WORKFLOW_ENDPOINT}/original_workflow`,
-            workflowRunId: "wfr_original_workflow",
-          }), 0)
-
-        },
-        responseFields: { body: "msgId", status: 200 },
-        receivesRequest: {
-          method: "POST",
-          url: `${MOCK_QSTASH_SERVER_URL}/v2/publish/${WORKFLOW_ENDPOINT}/${workflowId}`,
-          token,
-          body: {
-            body: "\"some-body\"",
-            headers: {
-              "Upstash-Workflow-Init": ["false"],
-              "Upstash-Workflow-RunId": ["wfr_original_workflow"],
-              "Upstash-Workflow-Url": ["https://requestcatcher.com/api/original_workflow"],
-              "Upstash-Feature-Set": ["LazyFetch,InitialBody"],
-              "Upstash-Workflow-Sdk-Version": ["1"],
-              "content-type": ["application/json"],
-              "Upstash-Telemetry-Sdk": ["sdk"],
-              "Upstash-Telemetry-Framework": ["framework"],
-              "Upstash-Telemetry-Runtime": ["runtime"],
-              "Upstash-Forward-Upstash-Workflow-Sdk-Version": ["1"],
-              "Upstash-Retries": ["3"],
-              "Upstash-Failure-Callback-Retries": ["3"],
-              "Upstash-Forward-original": ["original-headers-value"],
-              "Upstash-Failure-Callback-Forward-original": ["original-headers-value"],
-              "Upstash-Workflow-Runid": ["wfr_original_workflow"],
-            },
-            workflowRunId: "some-run-id",
-            workflowUrl: "https://requestcatcher.com/api/original_workflow",
-            step: {
-              stepId: 4,
-              concurrent: 1,
-              stepName: "invoke-step",
-              stepType: "Invoke",
-            },
-          },
-          headers: {
-            "upstash-retries": "6",
-            [`Upstash-Forward-${WORKFLOW_INVOKE_COUNT_HEADER}`]: "1"
-          },
-        },
-      })
     })
   })
